+++ src/PikaBot.h
@@ -26,33 +26,33 @@ class PikaBot
 {
 public:
     PikaBot(void);
-    uint16_t tempo = 120;
-    void calibrateMotors(uint8_t speed, uint16_t delay, uint8_t distance);
+    int tempo = 120;
+    void calibrateMotors(int speed, int delay, int distance);
     bool detectLine(IR sensor);
     bool isPressed();
-    void lineFollow(uint8_t speed);
-    void move(int16_t leftSpeed, int16_t rightSpeed);
-    void moveBackward(uint8_t speed);
-    void moveBackward(uint8_t speed, uint8_t distance);
-    void moveForward(uint8_t speed);
-    void moveForward(uint8_t speed, uint8_t distance);
+    void lineFollow(int speed);
+    void move(int leftSpeed, int rightSpeed);
+    void moveBackward(int speed);
+    void moveBackward(int speed, int distance);
+    void moveForward(int speed);
+    void moveForward(int speed, int distance);
     void play(String musicSheet);
-    void playTone(String pitch, uint8_t note);
-    void playToneMs(String pitch, unsigned long duration);
+    void playTone(String pitch, int note);
+    void playToneMs(String pitch, long duration);
     void resetTempo();
     void stop();
-    void turnLeft(uint8_t speed);
-    void turnLeft(uint8_t speed, uint8_t angle);
-    void turnRight(uint8_t speed);
-    void turnRight(uint8_t speed, uint8_t angle);
-    uint16_t ultrasonicDistance();
+    void turnLeft(int speed);
+    void turnLeft(int speed, int angle);
+    void turnRight(int speed);
+    void turnRight(int speed, int angle);
+    int ultrasonicDistance();
 
 private:
-    uint16_t _delayRef = 500;
-    uint8_t _distanceRef = 25;
-    uint8_t _speedRef = 255;
+    int _delayRef = 500;
+    int _distanceRef = 25;
+    int _speedRef = 255;
     unsigned long lastMotorInstructionTime = 0;
-    void _delayMotors(uint8_t speed, double distance);
+    void _delayMotors(int speed, double distance);
 };
 
 #endif
